refactor: create mocks.ts, add npm scripts
